requirements.txt
faiss-cpu==1.7.4
jq==1.4.1
lxml==4.9.3
openai==0.27.8
orjson==3.9.2
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.25.0
tiktoken==0.4.0
tqdm==4.65.0
